Extract methods of prebuilt rule assets client into separate files
