+++ x-pack/solutions/security/plugins/security_solution/common/api/detection_engine/prebuilt_rules/common/prebuilt_rule_assets_filter.ts
@@ -7,10 +7,8 @@
 
 import { z } from '@kbn/zod';
 
-export type ReviewPrebuiltRuleInstallationFilter = z.infer<
-  typeof ReviewPrebuiltRuleInstallationFilter
->;
-export const ReviewPrebuiltRuleInstallationFilter = z.object({
+export type PrebuiltRuleAssetsFilter = z.infer<typeof PrebuiltRuleAssetsFilter>;
+export const PrebuiltRuleAssetsFilter = z.object({
   fields: z.object({
     name: z
       .object({
